Feat(config): adjusting application build configuration
