--- config/webpack.dsv.config.js
@@ -1,56 +1,0 @@
-const path = require('path');
-const Webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const dotenv = require('dotenv');
-
-const env = dotenv.config().parsed;
-const envKeys = Object.keys(env).reduce((prev, next) => {
-  prev[`process.env.${next}`] = JSON.stringify(env[next]);
-  return prev;
-}, {});
-
-module.exports = {
-  entry: './src/index.js',
-  output: {
-    path: path.resolve(__dirname, '..', 'build'),
-    filename: 'index_bundle.js',
-    publicPath: '/',
-  },
-  module: {
-    rules: [
-      { test: /\.(js)$/, use: 'babel-loader' },
-      { test: /\.css$/, use: ['style-loader', 'css-loader'] },
-      {
-        test: /\.svg$/,
-        use: [
-          {
-            loader: 'babel-loader',
-          },
-          {
-            loader: 'url-loader',
-            options: {
-              jsx: true,
-            },
-          },
-        ],
-      },
-      {
-        test: /\.(png|jpg|gif)$/,
-        loader: 'url-loader',
-        query: {
-          limit: 8192,
-        },
-      },
-    ],
-  },
-  devServer: {
-    historyApiFallback: true,
-  },
-  mode: 'development',
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: 'dist/index.html',
-    }),
-    new Webpack.DefinePlugin(envKeys),
-  ],
-};
